Moved censoring UI into new file and started moving blocking contextual menu
